Refactor getline.ts files
