+++ getline.ts/src/index.ts
@@ -1,1 +1,2 @@
 export {Client} from './client'
+export {Loan} from './loan'
